Split layer io into a file and an impl file, added comments about what the differences between the functions are.
